requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.44.0
crispy-bootstrap4==2024.10
crispy-tailwind==1.0.3
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.3
django-allauth==65.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
waitress==3.0.2
whitenoise==6.8.2
